Core screens and scripts on screens

+++ StreamduckShared/Inputs/IInputDisplay.cs
@@ -11,16 +11,16 @@ public interface IInputDisplay {
 	 * Streamduck will hash its render structures and then call this to append the hash in case the devices requires
 	 * 2 different formats depending on which input is being rendered to
 	 */
-	int AppendHashKey(int key);
+	long AppendHashKey(long key);
 
 	/**
 	 * Lets the device process the image into format it needs,
 	 * key is derived from render structure and appended by the input
 	 */
-	Task UploadImage(int key, Image image);
+	Task UploadImage(long key, Image image);
 
 	/**
 	 * Should return true if image still exists, false if image was already deleted by the cache
 	 */
-	ValueTask<bool> ApplyImage(int key);
+	ValueTask<bool> ApplyImage(long key);
 }
